+++ tests/Feature/Repositories/UserRepository/FindAllWithTrashedTest.php
@@ -2,20 +2,24 @@
 
 declare(strict_types=1);
 
-namespace Tests\OldFeature\Repositories\UserRepository;
+namespace Tests\Feature\Repositories\UserRepository;
 
+use App\Models\User;
 use App\Repositories\UserRepository;
 use Illuminate\Support\Collection;
-use Tests\TestCase;
+use Tests\Feature\TestCase;
 
 class FindAllWithTrashedTest extends TestCase
 {
     private UserRepository $repository;
 
+    private User $user;
+
     protected function setUp(): void
     {
         parent::setUp();
         $this->repository = app(UserRepository::class);
+        $this->user = User::factory()->create();
     }
 
     public function test(): void
