+++ tests/Feature/Repositories/UserRepository/FindByEmailWithTrashedTest.php
@@ -2,20 +2,23 @@
 
 declare(strict_types=1);
 
-namespace Tests\OldFeature\Repositories\UserRepository;
+namespace Tests\Feature\Repositories\UserRepository;
 
 use App\Models\User;
 use App\Repositories\UserRepository;
-use Tests\TestCase;
+use Tests\Feature\TestCase;
 
 class FindByEmailWithTrashedTest extends TestCase
 {
+    private User $user;
+
     private UserRepository $repository;
 
     protected function setUp(): void
     {
         parent::setUp();
         $this->repository = app(UserRepository::class);
+        $this->user = User::factory()->create();
     }
 
     public function test(): void
